Melhorando estrutura do projeto
